% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.suffPCR}
\alias{coef.suffPCR}
\title{Method for returning coefficients of a fitted suffPCR object}
\usage{
\method{coef}{suffPCR}(object, i = NULL, d = NULL, ...)
}
\arguments{
\item{object}{a suffPCR object, created by \code{\link[=suffPCR]{suffPCR()}}}

\item{i}{optional vector of lambda indices to retrieve}

\item{d}{optional vector of d indices to retrieve}

\item{...}{ignored}
}
\value{
a (sparse) matrix of coefficients. Each column corresponds to one
lambda x d estimate
}
\description{
Method for returning coefficients of a fitted suffPCR object
}
