% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dlbcl}
\alias{dlbcl}
\title{Gene expression and patient clinical data for dlbcl}
\format{
A list with  named components:
\describe{
\item{x}{gene expression measurements for 240 patients on 7399 genes}
\item{gene_info}{a tibble containing descriptions of the genes}
\item{survivial_time}{time in years the patient survived, possibly right
sensored}
\item{survival_status}{a zero/one indicator of whether the patient died}
}
}
\source{
Andreas Rosenwald et al. The use of molecular profiling to predict
survival after chemotherapy for diffuse large-B-cell lymphoma.
\emph{New England Journal of Medicine}, 346(25):1937–1947, 2002.
}
\usage{
dlbcl
}
\description{
A dataset containing gene expression measurements, how long the patient
survived, whether they survived, and information about the genes for patients
with diffuse large-B-cell lymphoma. This data can be found in a variety of
R packages.
}
\keyword{datasets}
