% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsclc}
\alias{nsclc}
\title{Micro-RNA expression and patient data for nsclc}
\format{
A list with  named components:
\describe{
\item{x}{a named matrix with 123 patients and 939 columns containing gene
expression data. Column names are
patient ids to match the clinical data while row names provide
information about the measured miRNA.}
\item{clinical}{data frame of clinical data for the patients (birth date,
subtype, relapse information, chemo treatment, surgery, etc.)}
}
}
\source{
Vladimir Lazar et al. Integrated molecular portrait of non-small
cell lung cancers. \emph{BMC medical genomics}, 6(1):53, 2013.
}
\usage{
nsclc
}
\description{
Gene expression and patient clinical data for breast cancer
}
\details{
A dataset containing gene expression measurements along with a number of
clinical variables
}
\keyword{datasets}
