% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.suffpcr}
\alias{coef.suffpcr}
\title{Method for returning coefficients of a fitted suffpcr object}
\usage{
\method{coef}{suffpcr}(object, i = NULL, d = NULL, ...)
}
\arguments{
\item{object}{a suffpcr object, created by \code{\link[=suffpcr]{suffpcr()}}}

\item{i}{optional vector of lambda indices to retrieve}

\item{d}{optional vector of d indices to retrieve}

\item{...}{ignored}
}
\value{
a (sparse) matrix of coefficients. Each column corresponds to one
lambda x d estimate
}
\description{
Method for returning coefficients of a fitted suffpcr object
}
