% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffPCR.R
\name{suffPCR}
\alias{suffPCR}
\title{Sufficient principal component regression}
\usage{
suffPCR(
  X,
  Y,
  family = c("gaussian", "binomial"),
  d = 3,
  n_lambda = 10,
  maxnvar = ncol(X),
  lambda = NULL,
  lambda_max = NULL,
  lambda_min = NULL,
  lambda_seq = c("loglinear", "linear"),
  screening = TRUE
)
}
\arguments{
\item{family}{optional family argument to implement regression ("gaussian",
the default) or classification ("binomial")}

\item{d}{target PC dimension}

\item{n_lambda}{number of different lambda solutions to examine}

\item{maxnvar}{optional limit on the number of variables to consider}

\item{lambda}{optional vector of lambda values to use in the penalty}

\item{lambda_max}{optional largest value of lambda}

\item{lambda_min}{optional smallest value of lambda, must be non-negative
and less than lambda_max}

\item{lambda_seq}{should lambda be constructed on a loglinear or linear
scale between the minimum and maximum}

\item{screening}{do we screen as in algorithm 2}

\item{Xtrain}{n by p matrix of features}

\item{Ytrain}{length n response}
}
\value{

}
\description{
Estimates principal component regression (or classification) under the
assumption that the loadings are row-sparse. If this assumption is valid,
then the resulting predictors are guaranteed to be irrelevant for predicting
a response variable. This results in a sparse model with sparse linear
combinations of features used for the final prediction
}
\examples{
n <- 100
p <- 50
U <- rnorm(n)
V <- c(rnorm(5), rep(0, p - 5))
V <- V / sqrt(sum(V^2))
bstar <- V * (5 / (5.1))
X <- 5 * tcrossprod(U, V) + 0.1 * matrix(rnorm(n * p), n)
y <- U + rnorm(n)
out <- suffPCR(X, y, d=1:3)
}
